refactor: remove dead code flagged by the nitpicker audit
Remove unused exported API and their tests; where tests still exercise a live
internal path, keep a thin test-only wrapper (deadcode_wrappers_test.go) so
coverage holds without shipping dead code.

Removed: ErrorHandler and its helpers; wizard Validator/Detector and unused
exporter helper; unused validation.go validators and string helpers; unused
apperrors constructors; InitConfig, NewConfigurationLoaderWithOptions,
GetConfigurationSources/Enable/DisableSource, GenerateFromFile,
GetGitRepoRootAndInfo, and dead appconstants/embed/testutil helpers.
