+++ internal/config_helper_test.go
@@ -1,7 +1,6 @@
 package internal
 
 import (
-	"os"
 	"path/filepath"
 	"testing"
 
@@ -170,62 +169,6 @@ type ConfigHierarchySetup struct {
 	ActionFixture string // Fixture path for action config
 }
 
-// SetupConfigHierarchy creates a multi-level config hierarchy (global/repo/action).
-// Returns global config path, repo root, and action directory.
-//
-// Example:
-//
-//	globalPath, repoRoot, actionDir := SetupConfigHierarchy(t, tmpDir, ConfigHierarchySetup{
-//		GlobalFixture: testutil.TestConfigGlobalDefault,
-//		RepoFixture:   testutil.TestConfigRepoSimple,
-//		ActionFixture: testutil.TestConfigActionSimple,
-//	})
-func SetupConfigHierarchy(
-	t *testing.T,
-	baseDir string,
-	setup ConfigHierarchySetup,
-) (globalConfigPath, repoRoot, actionDir string) {
-	t.Helper()
-	// setupAndCreateConfigFixtures sets up config fixtures in a test directory.
-	// It creates the repo directory structure unconditionally and populates config files
-	// based on the provided setup.GlobalFixture, setup.RepoFixture, and
-	// setup.ActionFixture. Returns globalConfigPath, repoRoot, and actionDir.
-
-	// Create global config
-	if setup.GlobalFixture != "" {
-		globalConfigDir := filepath.Join(baseDir, testutil.TestDirDotConfig, testutil.TestBinaryName)
-		globalConfigPath = testutil.WriteFileInDir(
-			t, globalConfigDir, testutil.TestFileConfigYAML,
-			testutil.MustReadFixture(setup.GlobalFixture),
-		)
-	}
-
-	// Create repo config
-	repoRoot = filepath.Join(baseDir, testutil.ConfigFieldRepo)
-	if err := os.MkdirAll(repoRoot, 0o700); err != nil {
-		t.Fatalf("failed to create repo directory: %v", err)
-	}
-	if setup.RepoFixture != "" {
-		testutil.WriteFileInDir(
-			t, repoRoot, testutil.TestFileGHReadmeYAML,
-			testutil.MustReadFixture(setup.RepoFixture),
-		)
-	}
-
-	// Create action config
-	if setup.ActionFixture != "" {
-		actionDir = filepath.Join(repoRoot, testutil.ConfigFieldAction)
-		testutil.WriteFileInDir(
-			t, actionDir, testutil.TestFileConfigYAML,
-			testutil.MustReadFixture(setup.ActionFixture),
-		)
-	} else {
-		actionDir = repoRoot
-	}
-
-	return globalConfigPath, repoRoot, actionDir
-}
-
 // WriteConfigFixture writes a config fixture to a directory with standard config filename.
 // Returns the full path to the written config file.
 //
@@ -252,41 +195,6 @@ type ExpectedConfig struct {
 	Verbose      bool
 	Quiet        bool
 	GitHubToken  string
-}
-
-// AssertConfigFields asserts that config matches expected values for all non-empty fields.
-// Only checks fields that are set in expected (non-zero values).
-//
-// Example:
-//
-//	AssertConfigFields(t, config, ExpectedConfig{
-//		Theme:        testutil.TestThemeDefault,
-//		OutputFormat: "md",
-//		Verbose:      true,
-//	})
-func AssertConfigFields(t *testing.T, config *AppConfig, expected ExpectedConfig) {
-	t.Helper()
-	if expected.Theme != "" {
-		testutil.AssertEqual(t, expected.Theme, config.Theme)
-	}
-	if expected.OutputFormat != "" {
-		testutil.AssertEqual(t, expected.OutputFormat, config.OutputFormat)
-	}
-	if expected.OutputDir != "" {
-		testutil.AssertEqual(t, expected.OutputDir, config.OutputDir)
-	}
-	if expected.Template != "" {
-		testutil.AssertEqual(t, expected.Template, config.Template)
-	}
-	if expected.Schema != "" {
-		testutil.AssertEqual(t, expected.Schema, config.Schema)
-	}
-	// Always check booleans (they have meaningful zero values)
-	testutil.AssertEqual(t, expected.Verbose, config.Verbose)
-	testutil.AssertEqual(t, expected.Quiet, config.Quiet)
-	if expected.GitHubToken != "" {
-		testutil.AssertEqual(t, expected.GitHubToken, config.GitHubToken)
-	}
 }
 
 // assertGitHubClient validates GitHub client creation results.
