+++ internal/configuration_loader_helper_test.go
@@ -115,64 +115,3 @@ func checkThemeAndFormat(expectedTheme, expectedFormat string) func(t *testing.T
 		testutil.AssertEqual(t, expectedFormat, config.OutputFormat)
 	}
 }
-
-// AssertSourceEnabled fails the test if the specified source is not in the enabled sources list.
-// This helper reduces duplication in tests that verify configuration sources are enabled.
-//
-// Example:
-//
-//	AssertSourceEnabled(t, enabledSources, appconstants.ConfigSourceGlobal)
-func AssertSourceEnabled(
-	t *testing.T,
-	sources []appconstants.ConfigurationSource,
-	expectedSource appconstants.ConfigurationSource,
-) {
-	t.Helper()
-	for _, source := range sources {
-		if source == expectedSource {
-			return
-		}
-	}
-	t.Errorf("expected source %s to be enabled, but it was not found", expectedSource)
-}
-
-// AssertSourceDisabled fails the test if the specified source is in the enabled sources list.
-// This helper reduces duplication in tests that verify configuration sources are disabled.
-//
-// Example:
-//
-//	AssertSourceDisabled(t, enabledSources, appconstants.ConfigSourceGlobal)
-func AssertSourceDisabled(
-	t *testing.T,
-	sources []appconstants.ConfigurationSource,
-	expectedSource appconstants.ConfigurationSource,
-) {
-	t.Helper()
-	for _, source := range sources {
-		if source == expectedSource {
-			t.Errorf("expected source %s to be disabled, but it was found", expectedSource)
-
-			return
-		}
-	}
-}
-
-// AssertAllSourcesEnabled fails the test if any of the expected sources are not in the enabled sources list.
-// This helper reduces duplication in tests that verify multiple configuration sources are enabled.
-//
-// Example:
-//
-//	AssertAllSourcesEnabled(t, enabledSources,
-//	    appconstants.ConfigSourceGlobal,
-//	    appconstants.ConfigSourceRepo,
-//	    appconstants.ConfigSourceAction)
-func AssertAllSourcesEnabled(
-	t *testing.T,
-	sources []appconstants.ConfigurationSource,
-	expectedSources ...appconstants.ConfigurationSource,
-) {
-	t.Helper()
-	for _, expected := range expectedSources {
-		AssertSourceEnabled(t, sources, expected)
-	}
-}
