+++ internal/validation/strings_mutation_test.go
@@ -6,28 +6,6 @@ import (
 	"github.com/ivuorinen/gh-action-readme/testutil"
 )
 
-// urlTestCase represents a single URL parsing test case.
-type urlTestCase struct {
-	name        string
-	url         string
-	wantOrg     string
-	wantRepo    string
-	critical    bool
-	description string
-}
-
-// makeURLTestCase creates a URL test case with fewer lines of code.
-func makeURLTestCase(name, url, org, repo string, critical bool, desc string) urlTestCase {
-	return urlTestCase{
-		name:        name,
-		url:         url,
-		wantOrg:     org,
-		wantRepo:    repo,
-		critical:    critical,
-		description: desc,
-	}
-}
-
 // sanitizeTestCase represents a string sanitization test case.
 type sanitizeTestCase struct {
 	name        string
@@ -69,162 +47,6 @@ func makeFormatTestCase(name, org, repo, version, want string, critical bool, de
 		want:        want,
 		critical:    critical,
 		description: desc,
-	}
-}
-
-// TestParseGitHubURLMutationResistance tests URL parsing for regex and boundary mutations.
-// Critical mutations to catch:
-// - Pattern order changes (SSH vs HTTPS precedence)
-// - len(matches) >= 3 changed to > 3, == 3, etc.
-// - Return statement modifications (returning wrong indices).
-func TestParseGitHubURLMutationResistance(t *testing.T) {
-	tests := []urlTestCase{
-		// HTTPS URLs
-		makeURLTestCase(
-			"https_standard",
-			testutil.MutationURLHTTPS,
-			testutil.MutationOrgOctocat,
-			testutil.MutationRepoHelloWorld,
-			false,
-			"Standard HTTPS URL",
-		),
-		makeURLTestCase(
-			"https_with_git_extension",
-			testutil.MutationURLHTTPSGit,
-			testutil.MutationOrgOctocat,
-			testutil.MutationRepoHelloWorld,
-			true,
-			".git extension handled by (?:\\.git)? regex",
-		),
-
-		// SSH URLs
-		makeURLTestCase(
-			"ssh_standard",
-			testutil.MutationURLSSH,
-			testutil.MutationOrgOctocat,
-			testutil.MutationRepoHelloWorld,
-			true,
-			"SSH URL with colon separator ([:/] pattern)",
-		),
-		makeURLTestCase(
-			"ssh_with_git_extension",
-			testutil.MutationURLSSHGit,
-			testutil.MutationOrgOctocat,
-			testutil.MutationRepoHelloWorld,
-			true,
-			"SSH with .git",
-		),
-
-		// Simple format
-		makeURLTestCase(
-			"simple_org_repo",
-			testutil.MutationURLSimple,
-			testutil.MutationOrgOctocat,
-			testutil.MutationRepoHelloWorld,
-			true,
-			"Simple org/repo format (second pattern)",
-		),
-
-		// Edge cases with special characters
-		makeURLTestCase(
-			"org_with_dash",
-			testutil.MutationURLSetupNode,
-			testutil.MutationOrgActions,
-			testutil.MutationRepoSetupNode,
-			false,
-			"Hyphen in repo name",
-		),
-		makeURLTestCase("org_with_number", "org123/repo456", "org123", "repo456", false, "Numbers in org/repo"),
-
-		// Boundary: len(matches) >= 3
-		makeURLTestCase(
-			"exactly_3_matches",
-			"a/b",
-			"a",
-			"b",
-			true,
-			"Minimal valid: exactly 3 matches (full, org, repo)",
-		),
-
-		// Invalid URLs (should return empty)
-		makeURLTestCase(
-			"no_slash_invalid",
-			"octocatHello-World",
-			testutil.MutationStrEmpty,
-			testutil.MutationStrEmpty,
-			true,
-			"No slash separator",
-		),
-		makeURLTestCase(
-			"empty_string",
-			testutil.MutationStrEmpty,
-			testutil.MutationStrEmpty,
-			testutil.MutationStrEmpty,
-			true,
-			"Empty string",
-		),
-		makeURLTestCase(
-			"only_org",
-			"octocat/",
-			testutil.MutationStrEmpty,
-			testutil.MutationStrEmpty,
-			true,
-			"Trailing slash, no repo",
-		),
-		makeURLTestCase(
-			"only_repo",
-			"/Hello-World",
-			testutil.MutationStrEmpty,
-			testutil.MutationStrEmpty,
-			true,
-			"Leading slash, no org",
-		),
-
-		// Pattern precedence tests
-		makeURLTestCase(
-			"github_com_in_middle",
-			testutil.MutationURLGitHubReadme,
-			testutil.MutationOrgIvuorinen,
-			testutil.MutationRepoGhActionReadme,
-			false,
-			"First pattern should match",
-		),
-
-		// Regex capture group tests
-		makeURLTestCase(
-			"multiple_slashes",
-			"octocat/Hello-World/extra",
-			testutil.MutationStrEmpty,
-			testutil.MutationStrEmpty,
-			false,
-			"Extra path segments invalid for simple format",
-		),
-
-		// .git extension edge cases: dotted repo names are preserved and a single
-		// trailing ".git" suffix is stripped.
-		makeURLTestCase(
-			"double_git_extension",
-			"octocat/Hello-World.git.git",
-			testutil.MutationOrgOctocat,
-			testutil.MutationRepoHelloWorld+".git",
-			true,
-			"Dotted repo names preserved; one trailing .git suffix stripped",
-		),
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			gotOrg, gotRepo := ParseGitHubURL(tt.url)
-
-			if gotOrg != tt.wantOrg {
-				t.Errorf("ParseGitHubURL(%q) org = %q, want %q (description: %s)",
-					tt.url, gotOrg, tt.wantOrg, tt.description)
-			}
-			if gotRepo != tt.wantRepo {
-				t.Errorf("ParseGitHubURL(%q) repo = %q, want %q (description: %s)",
-					tt.url, gotRepo, tt.wantRepo, tt.description)
-			}
-		})
 	}
 }
 
@@ -326,138 +148,6 @@ func TestSanitizeActionNameMutationResistance(t *testing.T) {
 			got := SanitizeActionName(tt.input)
 			if got != tt.want {
 				t.Errorf("SanitizeActionName(%q) = %q, want %q (description: %s)",
-					tt.input, got, tt.want, tt.description)
-			}
-		})
-	}
-}
-
-// TestTrimAndNormalizeMutationResistance tests whitespace normalization.
-// Critical mutations to catch:
-// - Regex quantifier changes (\s+ to \s*, \s, etc.)
-// - TrimSpace removal or reordering
-// - ReplaceAllString to different methods.
-func TestTrimAndNormalizeMutationResistance(t *testing.T) {
-	tests := []sanitizeTestCase{
-		// Leading/trailing whitespace
-		makeSanitizeTestCase("leading_whitespace", "  hello", "hello", true, "TrimSpace removes leading"),
-		makeSanitizeTestCase("trailing_whitespace", "hello  ", "hello", true, "TrimSpace removes trailing"),
-		makeSanitizeTestCase("both_sides_whitespace", "  hello  ", "hello", true, "TrimSpace removes both sides"),
-
-		// Internal whitespace normalization
-		makeSanitizeTestCase(
-			"double_space",
-			testutil.ValidationHelloWorld,
-			testutil.ValidationHelloWorld,
-			true,
-			"Double space to single (\\s+ pattern)",
-		),
-		makeSanitizeTestCase(
-			"triple_space",
-			"hello   world",
-			testutil.ValidationHelloWorld,
-			true,
-			"Triple space to single",
-		),
-		makeSanitizeTestCase(
-			"many_spaces",
-			"hello          world",
-			testutil.ValidationHelloWorld,
-			true,
-			"Many spaces to single (+ quantifier)",
-		),
-
-		// Mixed whitespace types
-		makeSanitizeTestCase(
-			"tab_character",
-			"hello\tworld",
-			testutil.ValidationHelloWorld,
-			true,
-			"Tab normalized to space (\\s includes tabs)",
-		),
-		makeSanitizeTestCase(
-			"newline_character",
-			"hello\nworld",
-			testutil.ValidationHelloWorld,
-			true,
-			"Newline normalized to space (\\s includes newlines)",
-		),
-		makeSanitizeTestCase(
-			"carriage_return",
-			"hello\rworld",
-			testutil.ValidationHelloWorld,
-			true,
-			"CR normalized to space",
-		),
-		makeSanitizeTestCase(
-			"mixed_whitespace",
-			"hello \t\n world",
-			testutil.ValidationHelloWorld,
-			true,
-			"Mixed whitespace types to single space",
-		),
-
-		// Combined leading/trailing and internal
-		makeSanitizeTestCase(
-			"all_whitespace_issues",
-			"  hello   world  ",
-			testutil.ValidationHelloWorld,
-			true,
-			"Trim + normalize internal",
-		),
-
-		// Edge cases
-		makeSanitizeTestCase(
-			"empty_string",
-			testutil.MutationStrEmpty,
-			testutil.MutationStrEmpty,
-			true,
-			testutil.MutationDescEmptyInput,
-		),
-		makeSanitizeTestCase("only_spaces", "     ", testutil.MutationStrEmpty, true, "Only spaces (trimmed to empty)"),
-		makeSanitizeTestCase(
-			"only_whitespace_mixed",
-			" \t\n\r ",
-			testutil.MutationStrEmpty,
-			true,
-			"Only various whitespace types",
-		),
-		makeSanitizeTestCase("no_whitespace", "hello", "hello", false, "No whitespace to normalize"),
-		makeSanitizeTestCase(
-			"single_space_valid",
-			testutil.ValidationHelloWorld,
-			testutil.ValidationHelloWorld,
-			false,
-			"Already normalized",
-		),
-
-		// Multiple words
-		makeSanitizeTestCase(
-			"three_words_excess_spaces",
-			"one   two   three",
-			"one two three",
-			false,
-			"Three words with excess spaces",
-		),
-
-		// Unicode whitespace
-		makeSanitizeTestCase(
-			"regular_space",
-			testutil.ValidationHelloWorld,
-			testutil.ValidationHelloWorld,
-			false,
-			"Regular ASCII space",
-		),
-
-		// Quantifier verification (\s+ means one or more)
-		makeSanitizeTestCase("single_space_between", "a b", "a b", true, "Single space not collapsed (need + for >1)"),
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := TrimAndNormalize(tt.input)
-			if got != tt.want {
-				t.Errorf("TrimAndNormalize(%q) = %q, want %q (description: %s)",
 					tt.input, got, tt.want, tt.description)
 			}
 		})
@@ -654,80 +344,6 @@ func TestFormatUsesStatementMutationResistance(t *testing.T) {
 			if got != tt.want {
 				t.Errorf("FormatUsesStatement(%q, %q, %q) = %q, want %q (description: %s)",
 					tt.org, tt.repo, tt.version, got, tt.want, tt.description)
-			}
-		})
-	}
-}
-
-// TestCleanVersionStringMutationResistance tests version cleaning for operation order.
-// Critical mutations to catch:
-// - TrimSpace removal
-// - TrimPrefix removal or wrong prefix
-// - Operation order (trim then prefix vs prefix then trim).
-func TestCleanVersionStringMutationResistance(t *testing.T) {
-	tests := []sanitizeTestCase{
-		// v prefix removal
-		makeSanitizeTestCase(
-			"v_prefix_removed",
-			testutil.TestVersionSemantic,
-			"1.2.3",
-			true,
-			"TrimPrefix(\"v\") applied",
-		),
-		makeSanitizeTestCase("no_v_prefix_unchanged", "1.2.3", "1.2.3", true, "No v prefix to remove"),
-
-		// Whitespace handling
-		makeSanitizeTestCase("leading_whitespace", "  v1.2.3", "1.2.3", true, "TrimSpace before TrimPrefix"),
-		makeSanitizeTestCase("trailing_whitespace", "v1.2.3  ", "1.2.3", true, "TrimSpace applied"),
-		makeSanitizeTestCase("both_whitespace_and_v", "  v1.2.3  ", "1.2.3", true, "Both TrimSpace and TrimPrefix"),
-
-		// Operation order critical
-		makeSanitizeTestCase(
-			"whitespace_before_v",
-			" v1.2.3",
-			"1.2.3",
-			true,
-			"TrimSpace must happen before TrimPrefix",
-		),
-
-		// Edge cases
-		makeSanitizeTestCase("only_v", "v", testutil.MutationStrEmpty, true, "Just v becomes empty"),
-		makeSanitizeTestCase(
-			"empty_string",
-			testutil.MutationStrEmpty,
-			testutil.MutationStrEmpty,
-			true,
-			testutil.MutationDescEmptyInput,
-		),
-		makeSanitizeTestCase("only_whitespace", "   ", testutil.MutationStrEmpty, true, "Only spaces"),
-
-		// Multiple v's
-		makeSanitizeTestCase(
-			"double_v",
-			"vv1.2.3",
-			testutil.TestVersionSemantic,
-			true,
-			"Only first v removed (TrimPrefix, not ReplaceAll)",
-		),
-
-		// No changes needed
-		makeSanitizeTestCase("already_clean", "1.2.3", "1.2.3", false, "Already clean"),
-
-		// Real-world versions
-		makeSanitizeTestCase("semver_with_v", testutil.MutationVersionV2, "2.5.1", false, "Realistic semver"),
-		makeSanitizeTestCase("semver_no_v", "2.5.1", "2.5.1", false, "Realistic semver without v"),
-
-		// Whitespace variations
-		makeSanitizeTestCase("tab_character", "\tv1.2.3", "1.2.3", true, "Tab handled by TrimSpace"),
-		makeSanitizeTestCase("newline", "v1.2.3\n", "1.2.3", true, "Newline handled by TrimSpace"),
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := CleanVersionString(tt.input)
-			if got != tt.want {
-				t.Errorf("CleanVersionString(%q) = %q, want %q (description: %s)",
-					tt.input, got, tt.want, tt.description)
 			}
 		})
 	}
