+++ internal/validation/strings_property_test.go
@@ -122,191 +122,6 @@ func registerUsesStatementVersionPrefixProperty(properties *gopter.Properties) {
 	)
 }
 
-// TestStringNormalizationProperties verifies idempotency and whitespace properties.
-func TestStringNormalizationProperties(t *testing.T) {
-	properties := gopter.NewProperties(nil)
-	registerStringNormalizationProperties(properties)
-	properties.TestingRun(t)
-}
-
-func registerStringNormalizationProperties(properties *gopter.Properties) {
-	// Property 1: Idempotency - normalizing twice produces same result as once
-	properties.Property("normalization is idempotent",
-		prop.ForAll(
-			func(input string) bool {
-				n1 := TrimAndNormalize(input)
-				n2 := TrimAndNormalize(n1)
-
-				return n1 == n2
-			},
-			gen.AnyString(),
-		),
-	)
-
-	// Property 2: No consecutive spaces in output
-	properties.Property("normalized string has no consecutive spaces",
-		prop.ForAll(
-			func(input string) bool {
-				result := TrimAndNormalize(input)
-
-				return !strings.Contains(result, "  ")
-			},
-			gen.AnyString(),
-		),
-	)
-
-	// Property 3: No leading whitespace
-	properties.Property("normalized string has no leading whitespace",
-		prop.ForAll(
-			func(input string) bool {
-				result := TrimAndNormalize(input)
-
-				if result == "" {
-					return true
-				}
-
-				return !strings.HasPrefix(result, " ") &&
-					!strings.HasPrefix(result, "\t") &&
-					!strings.HasPrefix(result, "\n")
-			},
-			gen.AnyString(),
-		),
-	)
-
-	// Property 4: No trailing whitespace
-	properties.Property("normalized string has no trailing whitespace",
-		prop.ForAll(
-			func(input string) bool {
-				result := TrimAndNormalize(input)
-
-				if result == "" {
-					return true
-				}
-
-				return !strings.HasSuffix(result, " ") &&
-					!strings.HasSuffix(result, "\t") &&
-					!strings.HasSuffix(result, "\n")
-			},
-			gen.AnyString(),
-		),
-	)
-
-	// Property 5: All-whitespace input becomes empty
-	properties.Property("whitespace-only input becomes empty",
-		prop.ForAll(
-			func() bool {
-				// Generate whitespace-only strings
-				whitespaceOnly := "   \t\n\r   "
-				result := TrimAndNormalize(whitespaceOnly)
-
-				return result == ""
-			},
-		),
-	)
-}
-
-// TestVersionCleaningProperties verifies version string cleaning properties.
-// versionCleaningIdempotentProperty verifies cleaning twice produces same result.
-func versionCleaningIdempotentProperty(version string) bool {
-	v1 := CleanVersionString(version)
-	v2 := CleanVersionString(v1)
-
-	return v1 == v2
-}
-
-// versionRemovesSingleVProperty verifies single 'v' is removed.
-func versionRemovesSingleVProperty(version string) bool {
-	result := CleanVersionString(version)
-	if result == "" {
-		return true
-	}
-
-	trimmed := strings.TrimSpace(version)
-	if strings.HasPrefix(trimmed, "v") && !strings.HasPrefix(trimmed, "vv") {
-		return !strings.HasPrefix(result, "v")
-	}
-
-	return true
-}
-
-// versionHasNoBoundaryWhitespaceProperty verifies no leading/trailing whitespace.
-func versionHasNoBoundaryWhitespaceProperty(version string) bool {
-	result := CleanVersionString(version)
-	if result == "" {
-		return true
-	}
-
-	return !strings.HasPrefix(result, " ") &&
-		!strings.HasSuffix(result, " ") &&
-		!strings.HasPrefix(result, "\t") &&
-		!strings.HasSuffix(result, "\t")
-}
-
-// whitespaceOnlyVersionBecomesEmptyProperty verifies whitespace-only inputs become empty.
-func whitespaceOnlyVersionBecomesEmptyProperty() bool {
-	whitespaceInputs := []string{"   ", "\t\t", "\n", " \t\n "}
-	for _, input := range whitespaceInputs {
-		result := CleanVersionString(input)
-		if result != "" {
-			return false
-		}
-	}
-
-	return true
-}
-
-// nonVContentPreservedProperty verifies non-v content is preserved and trimmed.
-func nonVContentPreservedProperty(content string) bool {
-	trimmed := strings.TrimSpace(content)
-	if trimmed == "" || strings.HasPrefix(trimmed, "v") {
-		return true // Skip these cases
-	}
-
-	result := CleanVersionString(content)
-
-	return result == trimmed
-}
-
-func TestVersionCleaningProperties(t *testing.T) {
-	properties := gopter.NewProperties(nil)
-
-	// Property 1: Idempotency - cleaning twice produces same result
-	properties.Property("version cleaning is idempotent",
-		prop.ForAll(versionCleaningIdempotentProperty, gen.AnyString()),
-	)
-
-	// Property 2: Result never starts with single 'v' (TrimPrefix removes only one)
-	properties.Property("cleaned version removes single leading v",
-		prop.ForAll(versionRemovesSingleVProperty, gen.AnyString()),
-	)
-
-	// Property 3: No leading/trailing whitespace in result
-	properties.Property("cleaned version has no boundary whitespace",
-		prop.ForAll(versionHasNoBoundaryWhitespaceProperty, gen.AnyString()),
-	)
-
-	// Property 4: Whitespace-only input becomes empty
-	properties.Property("whitespace-only version becomes empty",
-		prop.ForAll(whitespaceOnlyVersionBecomesEmptyProperty),
-	)
-
-	// Property 5: Preserves non-v content and trims whitespace
-	properties.Property("non-v content is preserved",
-		prop.ForAll(
-			nonVContentPreservedProperty,
-			gen.OneGenOf(
-				gen.AlphaString(),
-				gen.AlphaString().Map(func(s string) string { return "  " + s }),
-				gen.AlphaString().Map(func(s string) string { return s + "  " }),
-				gen.AlphaString().Map(func(s string) string { return "  " + s + "  " }),
-				gen.AlphaString().Map(func(s string) string { return "\t" + s + "\n" }),
-			),
-		),
-	)
-
-	properties.TestingRun(t)
-}
-
 // TestSanitizeActionNameProperties verifies action name sanitization properties.
 func TestSanitizeActionNameProperties(t *testing.T) {
 	properties := gopter.NewProperties(nil)
@@ -392,100 +207,4 @@ func TestSanitizeActionNameProperties(t *testing.T) {
 	)
 
 	properties.TestingRun(t)
-}
-
-// TestParseGitHubURLProperties verifies URL parsing properties.
-func TestParseGitHubURLProperties(t *testing.T) {
-	properties := gopter.NewProperties(nil)
-	registerGitHubURLProperties(properties)
-	properties.TestingRun(t)
-}
-
-// registerGitHubURLProperties registers all GitHub URL parsing property tests.
-func registerGitHubURLProperties(properties *gopter.Properties) {
-	registerGitHubURLEmptyInputProperty(properties)
-	registerGitHubURLSimpleFormatProperty(properties)
-	registerGitHubURLNoSlashesProperty(properties)
-	registerGitHubURLInvalidInputProperty(properties)
-	registerGitHubURLConsistencyProperty(properties)
-}
-
-// registerGitHubURLEmptyInputProperty tests empty URL produces empty results.
-func registerGitHubURLEmptyInputProperty(properties *gopter.Properties) {
-	properties.Property("empty URL produces empty org and repo",
-		prop.ForAll(
-			func() bool {
-				org, repo := ParseGitHubURL("")
-
-				return org == "" && repo == ""
-			},
-		),
-	)
-}
-
-// registerGitHubURLSimpleFormatProperty tests simple org/repo format parsing.
-func registerGitHubURLSimpleFormatProperty(properties *gopter.Properties) {
-	properties.Property("simple org/repo format always parses correctly",
-		prop.ForAll(
-			func(org, repo string) bool {
-				if org == "" || repo == "" || strings.Contains(org, "/") || strings.Contains(repo, "/") {
-					return true
-				}
-				gotOrg, gotRepo := ParseGitHubURL(org + "/" + repo)
-
-				return gotOrg == org && gotRepo == repo
-			},
-			gen.AlphaString().SuchThat(func(s string) bool {
-				return len(s) > 0 && !strings.Contains(s, "/") && !strings.Contains(s, ".")
-			}),
-			gen.AlphaString().SuchThat(func(s string) bool {
-				return len(s) > 0 && !strings.Contains(s, "/")
-			}),
-		),
-	)
-}
-
-// registerGitHubURLNoSlashesProperty tests parsed results never contain slashes.
-func registerGitHubURLNoSlashesProperty(properties *gopter.Properties) {
-	properties.Property("parsed org and repo never contain slashes",
-		prop.ForAll(
-			func(url string) bool {
-				org, repo := ParseGitHubURL(url)
-
-				return !strings.Contains(org, "/") && !strings.Contains(repo, "/")
-			},
-			gen.AnyString(),
-		),
-	)
-}
-
-// registerGitHubURLInvalidInputProperty tests invalid URLs produce empty results.
-func registerGitHubURLInvalidInputProperty(properties *gopter.Properties) {
-	properties.Property("URLs without slash produce empty result",
-		prop.ForAll(
-			func(url string) bool {
-				if strings.Contains(url, "/") || strings.Contains(url, "github.com") {
-					return true
-				}
-				org, repo := ParseGitHubURL(url)
-
-				return org == "" && repo == ""
-			},
-			gen.AlphaString(),
-		),
-	)
-}
-
-// registerGitHubURLConsistencyProperty tests org and repo are both empty or both non-empty.
-func registerGitHubURLConsistencyProperty(properties *gopter.Properties) {
-	properties.Property("org and repo are both empty or both non-empty",
-		prop.ForAll(
-			func(url string) bool {
-				org, repo := ParseGitHubURL(url)
-
-				return (org == "" && repo == "") || (org != "" && repo != "")
-			},
-			gen.AnyString(),
-		),
-	)
 }
