--- internal/wizard/validator_test_helpers.go
@@ -1,1 +1,0 @@
-package wizard
